:construction: workout modal content refactor WIP
refactored notes, modal, and title. going to double back and write tests before moving on to exercises and exercise form
