+++ src/components/dash/views/week/Card.module.scss
