+++ src/components/dash/workoutmodal/Tags.module.scss
@@ -1,0 +1,32 @@
+@import '../../../styles/variables.scss';
+@import '../../../styles/mixins.scss';
+
+.container {
+  margin-left: 3rem;
+
+  .head {
+    font-size: 1.3rem;
+    padding-top: 1rem;
+    margin-bottom: 1.25rem;
+    color: $gray2;
+  }
+
+  .empty {
+    @extend .head;
+    font-size: 1.2rem;
+    margin: 0;
+  }
+
+  .tag {
+    min-width: 40px;
+    min-height: 30px;
+    border-radius: 4px;
+    padding: 0.75rem;
+    margin: 0.5rem 0.5rem 0.5rem 0;
+    color: white;
+    font-size: 1.3rem;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+}
