+++ src/components/dash/workoutmodal/WorkoutModal.module.scss
@@ -1,0 +1,14 @@
+@import '../../../styles/variables.scss';
+@import '../../../styles/mixins.scss';
+
+.content {
+  margin-left: 2rem;
+
+  @include small {
+    margin-left: 1rem;
+  }
+  .data {
+    width: 100%;
+    margin-right: 4rem;
+  }
+}
