+++ src/components/dash/workoutmodal/WorkoutModal.tsx
@@ -1,8 +1,13 @@
 import React, { memo } from 'react';
 import Modal from 'react-modal';
-import WorkoutTitle from './data/WorkoutTitle';
-import WorkoutContent from './data/WorkoutContent';
-import { useModalStyles } from '../workouts/modalStyles';
+import styles from './WorkoutModal.module.scss';
+import Flex from 'src/components/lib/Flex';
+import { useWindowSize } from 'react-use';
+import WorkoutOptions from './optionsmenu/WorkoutOptions';
+import Tags from './Tags';
+import WorkoutExercises from './data/exercises/WorkoutExercises';
+import Title from './Title';
+import Notes from './Notes';
 
 if (process.env.NODE_ENV !== 'test') Modal.setAppElement('#root');
 
@@ -12,20 +17,54 @@ interface Props {
   time: number;
 }
 
-// component for both adding a workout and viewing a saved workout
-// populates with appropriate content depending on global modal context (found in global store)
 const WorkoutModal: React.FC<Props> = ({ modal, closeModal, time }) => {
+  const { width } = useWindowSize();
+
+  const sizes = {
+    large: '750px',
+    largeMedium: '585px',
+    medium: '450px',
+    small: '300px'
+  };
+
+  const setModalSize = (): string => {
+    if (width <= 500) {
+      return sizes.small;
+    } else if (width <= 650) {
+      return sizes.medium;
+    } else if (width <= 800) {
+      return sizes.largeMedium;
+    } else {
+      return sizes.large;
+    }
+  };
+
   return (
     <Modal
+      style={{
+        content: {
+          width: setModalSize(),
+          margin: '0 auto',
+          background: '#f2f2f2',
+          border: 0,
+          padding: width <= 500 ? '10px' : '20px'
+        }
+      }}
       contentLabel='Add Workout Modal'
       isOpen={modal}
-      style={useModalStyles()}
       data-testid='modal'
     >
-      <section className='workout-modal-content'>
-        <WorkoutTitle closeModal={closeModal} />
-        <WorkoutContent closeModal={closeModal} time={time} />
-      </section>
+      <Flex fd='column' css={styles.content}>
+        <Title closeModal={closeModal} />
+        <Flex justify='space-between' fd={width <= 800 ? 'column' : undefined}>
+          <div className={styles.data}>
+            <Tags />
+            <Notes />
+            <WorkoutExercises />
+          </div>
+          <WorkoutOptions time={time} closeParentModal={closeModal} />
+        </Flex>
+      </Flex>
     </Modal>
   );
 };
