--- src/components/dash/workoutmodal/data/workouttags/tagstyles.ts
@@ -1,16 +1,0 @@
-// eslint-disable-next-line
-export const tagStyles = (color: string) => {
-  return {
-    background: color,
-    minWidth: '40px',
-    minHeight: '30px',
-    borderRadius: '4px',
-    padding: '0.75rem',
-    margin: '0.5rem 0.5rem 0.5rem 0',
-    color: 'white',
-    fontSize: '1.3rem',
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center'
-  };
-};
