+++ src/pages/Dashboard.tsx
@@ -12,9 +12,9 @@ import { fetchWorkoutsAction } from 'src/actions/fetchWorkoutsActions';
 import { fetchWorkoutsQuery } from 'src/utils/queries';
 import Controls from '../components/dash/controls/Controls';
 import { setHead, generateWeek, generateMonth } from 'src/utils/momentUtils';
-import Column from '../components/dash/workouts/week/Column';
+import Column from '../components/dash/views/week/Column';
 import WorkoutModal from '../components/dash/workoutmodal/WorkoutModal';
-import Cell from '../components/dash/workouts/month/Cell';
+import Cell from '../components/dash/views/month/Cell';
 import { Scope, Ctx } from 'src/types/Types';
 import SubNav from '../components/dash/subnav/SubNav';
 import { Helmet } from 'react-helmet-async';
