+++ src/styles/substyles/dash/workoutmodal/index.scss
@@ -1,7 +1,3 @@
-@import './modal.scss';
-@import './title.scss';
 @import './options.scss';
-@import './tags.scss';
 @import './exercises.scss';
-@import './notes.scss';
 @import './confirmdelete.scss';
