--- src/styles/substyles/dash/workoutmodal/modal.scss
@@ -1,28 +1,0 @@
-@import '../../../variables.scss';
-@import '../../../mixins.scss';
-
-.workout-modal-content {
-  display: flex;
-  flex-direction: column;
-  margin-left: 2rem;
-
-  @include small {
-    margin-left: 1rem;
-  }
-
-  .workout-body-container {
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
-    margin-top: 2rem;
-
-    @include medium {
-      flex-direction: column;
-    }
-
-    .workout-data {
-      width: 100%;
-      margin-right: 4rem;
-    }
-  }
-}
