--- src/styles/substyles/dash/workoutmodal/title.scss
@@ -1,59 +1,0 @@
-@import '../../../mixins.scss';
-
-.workout-modal-head {
-  @include flex-center;
-  justify-content: space-between;
-
-  .workout-modal-head-left {
-    @include flex-center;
-    width: 100%;
-
-    .workout-spotter-logo {
-      color: $primary;
-      font-size: 1.3rem;
-      margin-right: 1.3rem;
-
-      @include small {
-        font-size: 1.1rem;
-      }
-    }
-
-    .workout-title {
-      width: 100%;
-      font-size: 2rem;
-      color: $gray1;
-      border: 0;
-      padding: 0.75rem;
-      border-radius: $border;
-      background: $gray3;
-      -webkit-appearance: none;
-
-      @include small {
-        font-size: 1.7rem;
-      }
-
-      &:focus {
-        outline: 0;
-        box-shadow: 0 0 0 2px $primary;
-        background: white;
-      }
-    }
-  }
-
-  .workout-exit-modal {
-    font-size: 2rem;
-    cursor: pointer;
-    padding: 0.9rem;
-    border-radius: 50%;
-    margin-left: 2rem;
-    @include hover;
-
-    @include small {
-      font-size: 1.7rem;
-    }
-
-    &:hover {
-      background-color: darken($gray3, 10%);
-    }
-  }
-}
